move macros to own file
